requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
dnspython==2.3.0
Flask==2.2.5
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
pymongo==4.5.0
Werkzeug==2.2.3
python-dotenv==1.0.0